feat: add check for bridgeless in both packages (#7407)
## Summary

There are some issues like
[this](https://github.com/software-mansion/react-native-reanimated/issues/7367)
where it is not obvious what was the reason of exception. It turned out
to be having `bridge` which both `Reanimated` and `Worklets` do not
accept. For easier problem solving, I suggest adding an explicit check
for this scenario.

I also added compliance to JSI spec for `Worklets` module to be in line
with `Reanimated`.

## Test plan

Run the linked repro or manually enable bridge mode e.g.
[here](https://github.com/facebook/react-native/blob/7ff5db3c462c369a8210674ab53642c3f34e9c37/packages/react-native/Libraries/AppDelegate/RCTDefaultReactNativeFactoryDelegate.mm#L121)
to see that it does not work.