chore: unify processBoxShadow accross Reanimated and CSS (#7406)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

Before we had two processBoxShadow functions - in Reanimated (based on
implementation from React Native) and in CSS Animations (our simplified
function). For better managment and future maintance we decided to unify
those two by leaving our simplified function. I've extracted common
functions in `/src/common` folder and workletized them.



## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->