fix: add hermes debugger for all debug variants (#7428)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->
PR analogous to https://github.com/facebook/react-native/pull/50897
fixing `HERMES_ENABLE_DEBUGGER` not being added to configs that are not
strictly named `Debug`. Should fix
https://github.com/software-mansion/react-native-reanimated/issues/7412.
Thanks @tomekzaw and @piaskowyk for helping with it.

## Test plan

Run app from
https://github.com/software-mansion/react-native-reanimated/issues/7412
and see that it doesn't crash.