refactor(Worklets)(makeShareableClone): add makeShareableString (#7481)
## Summary
The first in a series of PRs that are intended to refactor the logic of
the `makeShareableClone` function.
The current implementation does not have a single source of truth, as we
have two type checking logics for the passed value, one on the `JS` side
and one on the `Cpp` side.
The idea of the refactor is to implement one source of truth on the `JS`
side and call the corresponding method on the `Cpp` side e.g:
```js
if (typeof value === ‘string’) {
  return global._makeShareableString(value);
}
```

In addition, in this PR I have rewritten the `Shareables` example to
make it more readable and easier to use.
## Test plan
`Shareables` example
<video
src="https://github.com/user-attachments/assets/5d61009d-e31b-44fc-b4fd-c21a8c417612"
/>