refactor(Worklets)(makeShareableClone): add makeShareableNumber, makeShareableBoolean, makeShareableBigInt (#7491)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
This PR adds few new functions : 
- `makeShareableNumber`
- `makeShareableBoolean`
- `makeShareableBigInt`

and is part of a series of PRs that aim to refactor the
`makeSharableClone` function ->
https://github.com/software-mansion/react-native-reanimated/pull/7481
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
`Shareables` example
<video
src="https://github.com/user-attachments/assets/9affdcf7-beae-416f-a7da-01fa7c48423b"
/>


<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->