refactor(Worklets): overhaul closure capturing algorithm (#7475)
## Summary

Improving the way we capture closures so it can be extended for the new
bundling methods.

The difference is that now we traverse the original AST instead of a
freshly generated one - it allows for further detection whether some
identifiers are imported from modules etc.

## Test plan

All tests pass with no changes, example app works flawlessly and Babel
runtime tests pass.