Move `WorkletRuntimeCollector::install` from Reanimated to Worklets (#7526)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR moves `WorkletRuntimeCollector::install` call from
`react-native-reanimated` platform-specific code to C++ codebase in
`react-native-worklets` since `WorkletRuntimeCollector` is still
required when `react-native-worklets` is installed without
`react-native-reanimated`. I've decided to call it from
`RNRuntimeWorkletDecorator::decorate` which is already platform-agnostic
so there's no code duplication.

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->