fix: Incorrect fill mode in alternate animations (#7563)
## Summary

This PR fixes incorrect fill mode being applied in alternate animations.
The previous implementation didn't take it into account that the every
second animation iteration in alternate animations is revered. As a
result, the animation was either treated like the animation with the
`normal` direction (for `alternate`) or `reverse` (for
`alternate-reverse`).

I removed the forwards fill mode getter and fixed the progress
calculation after resetting animation progress when its settings change.
As a result, if the animation is finished, I can just read the progress
from the progress provider, without assuming that it is always 1 for
finished animations as it might be 0 for alternate ones.

## Example recordings

When `animationDirection` is set to `alternate` or `alternate-reverse`
and the animation has the `forwards` fill mode, the appearance of the
box should change when the iteration count is changed between 1 and 2.
It didn't work in the old implementation.

| Before | After |
|-|-|
| <video
src="https://github.com/user-attachments/assets/0002dee1-58bc-474c-b3a2-021241937394"
/> | <video
src="https://github.com/user-attachments/assets/ab606c64-d12b-44a6-a294-b1d025db1a7a"
/> |