feat: Inject reanimated module proxy to reanimated shadow node (#7573)
## Summary

This PR adds a possibility to access the `ReanimatedModuleProxy` from
the `ReanimatedShadowNode`. It also fixes android linking issue to make
it possible to access all reanimated library stuff.

## Test plan

You can open any example in the example app and see that it renders
fine.

You can also uncomment the following line in the
`packages/react-native-reanimated/Common/NativeView/react/renderer/components/rnreanimated/ComponentDescriptors.h`:

```cpp
// LOG(INFO) << "We can access the proxy: " << proxy->getCssTimestamp();
```

and see that the proxy can be accessed.

## TODO

- [ ] Fix gesture handler integration - in the current implementation
gestures aren't attached to AnimatedComponents because the
`ReanimatedView` is not laid out (because of `display: contents`).