feat: Add a possibility to pass a CSS transition to the native view (#7439)
## Summary

This PR adds a possibility to pass CSS transition config and the static
JS style from the JS side to the native `ReanimatedView` component.

It also adds custom constructors to the `ReanimatedShadowNode` that will
be used to properly update state (storing `CSSTransition` and more in
the future) based on updated JS props.