feat: Scheduled operations chaining (#7461)
## Summary

The purpose of this PR is to add a simple management of time-based
operations. It will be useful for CSS transitions/animations scheduling,
taking delay and other things into account.

### Key assumptions

- operations management must be easy (without mutexes, multiple separate
steps),
- removal of state related to the previous CSS transition/animation
should be simple (no necessity to clear state in multiple places, remove
delayed operations, etc.),
- the time management logic should be encapsulated in an object that is
shared between CSS transition and CSS animations managers (no need to
write the same logic twice),
- operations must be automatically batched (e.g. we don't want to call
the `run` method on the transition as soon as the props on the view
change but we need to wait until the next frame and run all transitions
in a batch to ensure that they start with the same timestamp. The same
applies to the animations and the new approach should fix
[this](https://github.com/orgs/software-mansion/projects/35?pane=issue&itemId=95513423)
issue).

### Implementation details

#### Operations chaining

The new API supports operations chaining, which looks like this:

```cpp
Operation()
  .doOnce(...)
  .waitFor(...)
  .doWhile(...)
  .build();
```

For example, when the new CSS transition is triggered, we need to: 

(see this
[PR](https://github.com/software-mansion/react-native-reanimated/pull/7462)
for more usage details)

- call the `run` method on the `CSSTransition` object once at the
beginning to adjust progress providers and style interpolators to
modified properties that triggered a new transition (we can use
`.doOnce(...)`)
- if the transition is delayed, we have to wait until the delay passes
(we can use `.waitFor(...)` - it can be used with a callback function
that reads the current minimum delay from the `CSSTransition` instance;
if it returns 0, then there is no delay and the next chained operation
is executed),
- after starting the transition, we want to update its progress on every
frame (we can use `.doWhile(...)` - update the progress and return
`true` as long as the transition is running or return `false` when the
transition state changes to `Finished` to remove the operation from the
operations loop).

#### Simple cleanup

We need to do cleanup e.g. when the new transition is started before the
previous one finishes or the transition is removed from the view. In the
previous implementation we had to remove the transition from the active
transitions map and remove from the delayed items manager manually.

The new implementation makes it simpler by returning an
`OperationHandle` from the `schedule()` method called on the
`OperationsLoop` instance, which can be then used to cancel the
operations chain, no matter what is the current state (if the transition
is delayed/running).

```cpp
// Schedule operations
const auto handle = operationsLoop_->schedule(...);

// Cancel the operations chain later on
operationsLoop->remove(handle);
```

## Code example

```cpp
auto start = getAnimationTimestamp_();
auto op = Operation()
  .doOnce([start](double timestamp) {
    LOG(INFO) << "ReanimatedModuleProxy::ReanimatedModuleProxy() [1] doOnce: " << timestamp << " " << start;
  })
  .waitFor(2500.0)
  .doWhile([start](double timestamp) {
    LOG(INFO) << "ReanimatedModuleProxy::ReanimatedModuleProxy() [2] doWhile: " << timestamp << " " << start
              << " continue: " << (timestamp < start + 5000.0);
    return timestamp < start + 5000.0;
  })
  .waitFor([]() { return 2000.0; })
  .doOnce([start](double timestamp) {
    LOG(INFO) << "ReanimatedModuleProxy::ReanimatedModuleProxy() [3] doOnce: " << timestamp << " " << start;
  })
  .build();

const auto handle = operationsLoop_->schedule(std::move(op));
```