feat: New CSS managers on the JS side (#7467)
## Summary

This PR adds the implementation of new JS managers for CSS transition
and animations. They are much simpler than previous implementation
because we moved the props diffing logic to CPP as we just pass the
processed config via the `ReanimatedView` props and react to all changes
in CPP.

The `CSSAnimationsManager` is a bit more complex as it needs to register
CSS animation keyframes (via a JSI call) before they are used. This is
an optimization to allow reusing the same animation keyframes (thus the
CPP style interpolator) for multiple views without having to parse
keyframes and create separate interpolators multiple times.