fix: Reanimated Component Descriptor initialization on reloads (#7508)
## Summary

- fixes the initialization of the `ReanimatedViewComponentDescriptor`
(it used some strange initialization logic pulling the `proxy` instance
from the context container before, which didn't sometimes work. In
general, the `ReanimatedModuleProxy` is usually initialized first, but
it is not the case during hot reloads, when the order of initialization
of `ReanimatedViewComponentDescriptor` and `ReanimatedModuleProxy` can
be different. Because of that, we ended up with an uninitialized
`ReanimatedModuleProxy` instance with invalid shadow nodes without
properly initialized state object that depends on the `proxy`)