chore: Restore jsi::Value-based CSS config parsers (#7518)
## Summary

This PR restores changes introduced in the #7443 PR. I would like to
write a custom naive view props class with custom parsing for config
properties.

The old flow parsed `jsi::Values` to `folly::dynamic`, and then parsed
`folly::dynamic` to the config object.
In the new flow, props will be parsed from the `jsi::Value` type to the
config type directly which will result in a cleaner and simpler code.