feat(Worklets): new ShareableImport type (#7512)
## Summary

This pull request introduces new type of a Shareable, `ShareableImport`.
It's a type that, when extracted on another runtime, tries to "pull
itself" from the JavaScript bundle.

For example, imagine you have a function `foo` in a `module.js` module:

```js
// module.js
export function foo(){
    console.log('hello world');
}
```

If this function were to be serialized as `ShareableImport`, the
following code:

```js
import {runOnUI} from 'react-native-worklets';
import {foo} from 'module';

runOnUI(foo)();
```

Would actually "forward" the instruction `import {foo} from 'module'` to
the UI Runtime, and the UI Runtime would invoke (Metro) require to
obtain it and execute it.

For the time being there are no use cases for `ShareableImport` but they
will come in the upcoming PRs.

## Test plan

Nothing to test, it doesn't affect runtime at the moment.