fix: unitialized shared_ptr on android (#7576)
## Summary

In #7486 I made the function `createJSIWorkletsModuleProxy` which uses
uninitialized shared_ptr on first invocation - it was intentional, but
turns out that this uninitialized shared_ptr was the inteded `nullptr`
on iOS but a garbage memory on Android. I'm moving the line to the
constructor outside of initializer list to fix this.

## Test plan

I tested Android in both debug and release and it didn't crash anymore.