release: Reanimated 4.0.0-beta.5 (#7588)
## Summary

This PR is a release of the next beta version with a hot fix for CSS
animation crashes in `4.0.0-beta.4`.

This version is still compatible with react-native-worklets `0.3.0`.

I did a quick test, mainly mobile (iOS, Android) and quick tests for
web. I also tested the example from the #7582 PR which was used as a
repro for CSS crashes before.