chore: Add missing CSS properties helper warnings (#7616)
## Summary

This PR adds helper warnings for missing CSS animation/transition
properties.

### When warnings are shown?

In general, when the user might forgotten to specify one of essential
props and might struggle to identify why the animation/transition
doesn't work.

#### Animations

- `animationName` is missing but at least one other CSS animation prop
is specified
- `animationDuration` is not specified when the `animationName` prop
exists (quite often case that people assume that the default duration is
not 0)

#### Transitions

- `transitionDuration` is not specified when the `transitionProperty`
prop exists (similar to the issue with `animationDuration`)

## Examples

### Animations

![Screenshot 2025-06-03 at 17 42
50](https://github.com/user-attachments/assets/9109f18e-4179-4a02-9a00-9cbd40595c2e)

### Transitions

![Screenshot 2025-06-03 at 17 43
44](https://github.com/user-attachments/assets/a2c92ffc-8e64-4e35-940d-379d28db424f)