feat(Worklets): resolve global local variable names conflicts (#7577)
## Summary

Since the closure capturing mechanism got overhauled, we can actually
now only ignore global bindings if they are not explicitly bound in the
AST.

Thanks to this, the following code:

```ts
const window = [ ];

function foo() {
	'worklet';
	return window;
}
```

Would actually capture the `window` object. Previously it would be
ignored, since we didn't know if it referred to a bound variable or to
global `window` object.

I also fixed all parts of the code where `global.` accessor wasn't used
but should be used.

## Test plan

I ran the app and Jest tests with no regressions. I also added a few
tests to acknowledge the improvement in behavior.