feat(Worklets): workletizable modules (#7584)
## Summary

This pull request adds a stub of an API for 3rd party libraries that can
be used in Worklet Runtimes.

The premise is quite simple - if a library doesn't have any fatal side
effects for a Worklet Runtime, this primarily means that **a 3rd party
library used on Worklet Runtime cannot import anything from
`react-native`**. We might loosen these requirements in the future if we
implement some workarounds for these side effects.

The way to use a library on a Worklet Runtime is to register it as a
`workletizableModule` in `babel.config.ts`, i.e.

```js
// babel.config.js
/** @type {import('@babel/core').TransformOptions} */
module.exports = {
  plugins: [
    [
      'react-native-worklets/plugin',
      {
        experimentalBundling: true,
		workletizableModules: ['three'],
      },
    ],
  ],
};
```

This causes that imports in worklets from that library persist as
imports in worklets.

## Test plan

I ran Worklets Babel plugin Jest tests and there were no regressions.