fix(Worklets): remove unnecessary checks in makeShareableCloneOnUIRecursive (#7633)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
Remove unnecessary support for the `_makeShareableWorklet` and
`_makeShareableInitializer` functions
<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->

## Test plan
`WorkletRuntimes` example
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->