chore(Worklets): patches for experimental bundling (#7585)
## Summary

To execute the bundle on Worklet Runtimes we need to somehow access it.
Unfortunately, currently there are no APIs for that and we need to patch
React Native to expose them.

Once we settle for a good API both internally and externally I will
upstream these changes to
https://github.com/facebook/react-native/pull/50788#pullrequestreview-2869147501

## Test plan

Fabric example works on:
- [x] iOS debug
- [x] iOS prod
- [x] android debug
- [x] android prod