docs: fix Radon IDE banner (#7634)
In the production environment Docusaurus serves the HTML statically and
then hydrates the HTML shell with React. This caused the banner to
randomly pick the banner copy twice (different text was picked "on the
server" and then again "on the client") causing an ugly jump/rerender.

There was no way to catch that bug in dev mode as Docusaurus only runs
the page "client side" during development.

I've fixed the problem by only running the code only on the client side
and masking the hydration step with a slide and fade animation.

In the following recordings I'm repeatedly refreshing the page.

### Before



https://github.com/user-attachments/assets/7739e2f0-56d1-489d-b7ed-7442c1be4a6e


### After



https://github.com/user-attachments/assets/b5330d5f-2d7f-40f8-898d-a1c933c4cc1f