Refactor props allowlists (replace `addWhitelistedUIProps` and `addWhitelistedNativeProps` with `registerJSProps`) (#7642)
## Summary

This PR removes `addWhitelistedNativeProps` and `addWhitelistedUIProps`
functions from the public API of Reanimated as well as `configureProps`
and `adaptViewConfig` functions from Reanimated internals.

Thanks to this change, it will be possible to animate props of
`TextInput` component from react-native or `Circle` and `Path`
components from react-native-svg via `useAnimatedProps` without
explicitly registering props as native props.

As a reminder, "JS props" are props that need some preprocessing on the
JS side, for instance `G` component from react-native-svg. From now on,
by "native props" we mean all native props (from the spec) of a given
native component. "Native props" can be updated via ShadowTree.


https://github.com/user-attachments/assets/1f79bf3a-4433-4145-8b33-1cfc0432ba74

## Context

Reanimated 2 introduced `addWhitelistedNativeProps` and
`addWhitelistedUIProps` functions to register so-called "native props"
and "UI props". Native props had to be updated on the native modules
thread (and required layout recalculation) while UI props could be
updated directly on the views without layout recalculation. If some prop
wasn't registered as native or UI prop, it was referred to as a "JS
prop" because it was updated using `setNativeProps` scheduled on the JS
thread.

One of the limitations of `addWhitelistedNativeProps` and
`addWhitelistedUIProps` was the fact that the sets were shared for all
components, i.e. some prop name had to be treated either as a native or
a UI prop for all types of components. For instance, it was not possible
to register prop `x` as a native prop for `ComponentA` and at the same
time as a UI prop for `ComponentB`.

Reanimated 3 came with support for the New Architecture. First we
removed support for JS props on the New Architecture but then we
restored it later in
https://github.com/software-mansion/react-native-reanimated/pull/5009.
In this PR, we added `animatablePropNames_` set which was used to keep
the names both of native and UI props. If the prop name wasn't listed
there, it was treated as a JS prop.

In https://github.com/software-mansion/react-native-reanimated/pull/7014
we removed the distinction between native and UI props because of broken
hit boxes for transformed views caused by
`synchronouslyUpdatePropsOnUIThread` not committing the changes to the
ShadowTree. After that change, all props were committed to ShadowTree
(as an optimization, because `RawPropsParser` skips unknown props) and
props absent in `animatablePropNames_` were passed to `updateJSProps`
call.

Finally, this PR removes the need for manual registration of
"animatable" prop names via `addWhitelistedNativeProps` and
`addWhitelistedUIProps` so animating native props works automatically
out-of-the-box for custom native components. At the same time, this PR
introduces `registerJSProps` function which is used to define the names
of JS props for each native component separately. This way, we are able
to tell which props need to be treated as JS props in
`performOperations` as we can obtain the component name from a shadow
node using `shadowNode->getComponentName()` method. I also added a mutex
and lock because `registerJSProps` function is called from the RN
runtime and it writes to `jsPropsNamesForComponentNames_` unordered map
on the JS thread which is also read on the UI thread in
`performOperations`.

Note that we don't exclude JS props from prop updates when updating
native props via ShadowTree since `RawPropsParser` quietly skips those
so we can safely keep them and they will just get ignored.

## Changes

* Removed `configureProps` and introduced
`registerNativePropNamesForComponentName`
* Renamed `animatablePropNames_` to `nativePropNamesForComponentNames_`
* Renamed `filterNonAnimatableProps` to `filterNonNativeProps`
* Added synchronization for accessing
`nativePropNamesForComponentNames_` as it is accessed from the JS and UI
threads
* Added component name to JSI updates list in `AnimatedPropsRegistry`
* Replaced the implementation `addWhitelistedNativeProps` and
`addWhitelistedUIProps` with a warning
* Removed `addWhitelistedNativeProps` call in
`createAnimatedPropAdapter` and added a warning
* Removed `addWhitelistedNativeProps` calls in `PerformanceMonitor`
component as well as in the example app
* Removed unused `PropsAllowlists` and `assertNoOverlapInLists`

## TODO
* ~~Find a way to detect if given prop is a native or non-native prop
(in C++ or JS)~~
* ~~Decide if we keep the aforementioned functions for
backward-compatibility as no-ops or warnings or remove them completely~~
* ~~Add a possibility to override native props to be non-native (e.g.
some developer wants the `text` prop to be a JS prop)~~
* Find out what's the reason for this patch in the New Expensify app:
https://github.com/Expensify/App/blob/main/patches/react-native-reanimated%2B3.17.1%2B002%2BdontWhitelistTextProp.patch
* Fix `~TextInputState` crash on reload on iOS &rarr; reported as
https://github.com/software-mansion/react-native-reanimated/issues/7647

## Test plan