feat: Add CellRendererComponentStyle to Animated.FlatList (#6573)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary
This PR adds a new prop to `Animated.FlatList`,
`CellRendererComponentStyle`. It has a very similar description as
`react-native`'s `Pressable` style prop, where it can either be a
regular animated style, or a function that receives the index and the
corresponding item. The naming convention follows `react-native`'s
`FlatList` naming convention (e.g. `ListHeaderComponent` and
`ListHeaderComponentStyle`).

<!-- Explain the motivation for this PR. Include "Fixes #<number>" if
applicable. -->
The idea sprung when trying to implement a draggable `FlatList` with
`reanimated`'s `Animated.FlatList` and being unable to change the
`zIndex` of the item being dragged, resulting in it always being
rendered behind items of a higher index.

After the `CellRendererComponent` prop was removed from
`Animated.Flatlist`, it became impossible to control styles that need to
be altered in the top-level element - specifically `zIndex`.
While it would be possible to use `Animated.createAnimatedComponent` and
mimic the logic of `Animated.FlatList` for anything but the style of the
`CellRendererComponent`, this API seems to have at least [some
attraction](https://github.com/software-mansion/react-native-reanimated/issues/5931#issuecomment-2084410286)
within the community and therefore is being suggested.


## Test plan
<!-- Provide a minimal but complete code snippet that can be used to
test out this change along with instructions how to run it and a
description of the expected behavior. -->

Simply adding this code snippet as a property on an `Animated.FlatList`
should show that it functions as intended. See video below where this
snippet was added to the `InvertedFlatListExample.tsx` example.

```tsx
      CellRendererComponentStyle={({ index }) => ({
        backgroundColor: index % 2 === 0 ? 'red' : 'blue',
      })}
```


https://github.com/user-attachments/assets/76fa291d-f187-4d49-9277-742e41cb0167

Co-authored-by: Gunnsteinn <gunnsteinn@smittendating.com>
Co-authored-by: Krzysztof Piaskowy <krzysztof.piaskowy@swmansion.com>