chore: update circular dependency check (#7687)
## Summary
Together with @tjzel we agree that the circular dependency check is too
restrictive, moreover having circular dependency in typescript is not
forbidden, sometimes it is useful to simplify the code.
This PR restricts the circular dependency check only to `js` files in
the lib folder for `reanimated` and `worklets` packages.

## Test plan
🚀