chore(Worklets): detect runtime initialization failures with experimental bundling (#7638)
## Summary

Throw a QoL error when something went wrong and we received a different
error than expected during Worklet Runtime initialization when using
experimental bundling.

I had to move `reportFatalErrorOnJS` to C++ to be able to invoke it on a
wrong error. The changes are backwards compatible with legacy bundling.

## Test plan

Enable experimental bundling as per instructions in
`packages/react-native-worklets/experimentalBundling.md` and see that
there are no crashes.

If a Worklet Runtime failed to initialize, you'll get the following
error in JS now (I modified `workletRuntimeEntry.ts` to achieve those
configurations of errors):

| Initialization failure of UI Runtime (sync invocation) |
Initialization failure of Worklet Runtime (sync invocation) |
Initialization failure of Worklet Runtime (spawned from the UI Thread,
async invocation) |
|:--------:|:--------:|:--------:|
| ![Screenshot 2025-06-16 at 19 48
33](https://github.com/user-attachments/assets/124914ad-4571-4a9c-ae95-1ced7327a005)
| ![Screenshot 2025-06-16 at 19 50
05](https://github.com/user-attachments/assets/0db543e8-5cd5-4ffb-8441-9764981e959f)
| ![Screenshot 2025-06-16 at 20 00
53](https://github.com/user-attachments/assets/9b9e3361-6ba2-411a-9ccf-77ad91c20298)
|


For legacy bundling, throw an error in EmptyExample and see that it
works just as before.