fix(Worklets): reassigning runOnUIQueue (#7714)
## Summary

After

- #7365

outside variables inside worklets are no longer coming from closure but
instead from the factory parameters. Because of that, patterns like
this:

```js
let a = 0;

function foo(){
	'worklet';
	a = {};
}

function bar(){
	'worklet';
	console.log(a);
}

foo();
bar(); // should log '{}'
```

won't work, because modifying `a` modifies the parameter (local
reference) inside the worklet and we log `0` instead.

However, this is a highly discouraged anti-pattern and shouldn't be
practiced. Nevertheless, there's a place in our code where we use it -
for the queue of worklets to execute with `runOnUI`. The modification
there happens only on the RN runtime, so the workaround for it is not to
make `runOnUI` and `runOnUIAsync` worklets. This is a temporary fix, I
will need to rethink this approach with experimental bundling.

## Test plan

`runOnUIAsync` example works again after this PR.