chore: Migrate from ts-prune to knip (#7721)
## Summary

This PR replaces the old `ts-prune` unused code checking library by the
more modern equivalent - `knip`. `ts-prune` is not maintained for 4
years and I decided to use `knip`, because it is more up-to-date and
offers better customization. `knip` also marks code as unused more
properly and spots more unused code parts that can be removed, compared
to the `ts-prune`.