chore: Change updateKeyframes to work with folly::dynamic instead of jsi::Value (#7734)
## Summary

This PR is necessary to add support for component-specific
interpolators. I would have to store keyframes config object as a
`folly::dynamic` instead of creating the style interpolator right away,
because I cannot determine on which component type the animation will be
attached.

Only after attaching the animation to the view, I can pick the right
interpolators config object and create the style interpolator from the
persisted keyframes (I cannot persist a `jsi::Value`, so that's why I
need this change to `folly::dynamic`).

This PR is a part of PRs set that will add support for CSS animations
used on `react-native-svg` components.