Fix reparenting (#7732)
<!-- Thanks for submitting a pull request! We appreciate you spending
the time to work on these changes. Please follow the template so that
the reviewers can easily understand what the code changes affect. -->

## Summary

This PR fixes how we manage moving views (through reparenting) during a
layout animation. Our current behavior would not change the parent in
our animation state, which would cause us to try to apply updates with a
stale parent. This would cause a crash, because the view state for the
parent would be unavailable.

Hopefully closes #7493 

## Test plan

Check `[LA] Reparenting` example on Android. It should lead to a soft
crash in our example app. (In previous RN versions it would be a hard
crash)

---------

Co-authored-by: Tomasz Zawadzki <tomekzawadzki98@gmail.com>